+++ test/uploaded_file_test.rb
@@ -126,7 +126,7 @@ describe Shrine::UploadedFile do
     end
 
     it "does not include query params from shrine-url ids in extension" do
-      uploaded_file =uploaded_file("id" => "http://example.com/path.html?key=value", "storage" => "cache")
+      uploaded_file = uploaded_file("id" => "http://example.com/path.html?key=value", "storage" => "cache")
       assert_equal "html", uploaded_file.extension
 
       uploaded_file = uploaded_file("id" => "http://example.com/path?key=value", "storage" => "cache")
@@ -143,7 +143,6 @@ describe Shrine::UploadedFile do
       uploaded_file = uploaded_file("id" => "foo.xx?")
       assert_equal "xx?", uploaded_file.extension
     end
-
   end
 
   describe "#size" do